[ews-build.webkit.org] Use TwistedAdditions in place of Twisted requests https://bugs.webkit.org/show_bug.cgi?id=249663 rdar://103564994
Reviewed by Dewei Zhu.

* Tools/CISupport/ews-build/events.py:
(Events):
(Events.sendDataToEWS):
(Events.sendDataToGitHub):
(JSONProducer): Moved to twisted_additions.
* Tools/CISupport/ews-build/twisted_additions.py:
(TwistedAdditions.JSONProducer): Moved from events.py.
(TwistedAdditions.request): Allow caller to POST data.

Canonical link: https://commits.webkit.org/258542@main
